Change FrameData point vars to double (relative percentage of image)
Originally the FrameData point x's and y's were longs which held the
differences in pixels from the top left corner. It's been changed now to
be doubles which contain a 0 to 1 value relative to the image size:
X
	0 = left edge
	1 = right edge
Y
	0 = top edge
	1 = bottom edge

Essentially I'm getting a higher precision.

+++ src/FrameData.h
@@ -15,13 +15,11 @@
 #ifndef FRAMEDATA_H
 #define FRAMEDATA_H
 
-#include <vector>
-
 namespace Stringless {
     
 struct PointPair {
-    long x;
-    long y;
+    double x;
+    double y;
 };
 
 struct FrameData {
